add destroy and get on for fault

+++ app/Http/Controllers/FaultController.php
@@ -17,6 +17,12 @@ class FaultController extends Controller
         $data=Fault::all();
         return response()->json($data,200);
     }
+    public function show($id)
+    {
+        $data=Fault::find($id);
+        return response()->json($data,201);
+    }
+
 
     public function store()
     {        
@@ -31,11 +37,15 @@ class FaultController extends Controller
 
     public function destroy($id)
     {   // return response()->json($id,200);    
-        $id=Fault::find($id);
-        //return response()->json($id,200);
-
-        $isdel= $id->delete();
-        return response()->json($isdel,200);
+        $id=Fault::find($id);        
+        if(isset($id))
+        {           
+            $isdel= $id->delete();
+            return response()->json($isdel,200);
+        }
+        else
+            return response()->json(false,404);
+        
     }
     public static function Validation($id=0)
     {  
@@ -53,12 +63,12 @@ class FaultController extends Controller
             {
                 $fault= Fault::where('id',$id) ;              
                 $data=$fault->update($data);
-                return response()->json($data,201);
+                return response()->json($data,202);
             }
             else
             {                
                 $data=Fault::create($data);
-                return response()->json($data,200);
+                return response()->json($data,201);
             }
